I'm so proud of myself ❤
Its been so hard for me to manage my priorities towards the end of 2023. I got on this swiftly and got my certification after 4 hours on begin the course in one day. And I'm just grateful to my parents for always being my pillars❤. This isn't too big of a certification and it's definitely not my first😅, But I'm quite emotional this evening and since I feel my dreams will come true super soon, I want this to be a throwback.